Fixed cast pb in kahan_sum.hpp detected by Jenkins on MinGW64-4.9.0

+++ ssc/numeric/kahan_sum.hpp
@@ -53,7 +53,7 @@ namespace ssc
         {
             return kahan<typename boost::counting_iterator<long>,
                              Iterator,
-                             T>(boost::counting_iterator<long>(0),boost::counting_iterator<long>(std::distance(begin,end)), begin);
+                             T>(boost::counting_iterator<long>(0),boost::counting_iterator<long>((long)std::distance(begin,end)), begin);
         }
 
         template <typename T>  T kahan(const std::vector<T>& points)
